+++ yellowbox_snowglobe/session.py
@@ -97,12 +97,19 @@ class SnowGlobeSession:
     def _do_ignore(self, query: str) -> QUERY_RESPONSE:
         return None
 
+    def _restart_transaction(self) -> None:
+        self._transaction = self.connection.begin()
+
     def _do_commit(self, query: str) -> QUERY_RESPONSE:
-        self.transaction.commit()
+        if self.transaction.is_active:
+            self.transaction.commit()
+        self._restart_transaction()
         return None
 
     def _do_rollback(self, query: str) -> QUERY_RESPONSE:
-        self.transaction.rollback()
+        if self.transaction.is_active:
+            self.transaction.rollback()
+        self._restart_transaction()
         return None
 
     def _do_use_database(self, query: str) -> QUERY_RESPONSE:
@@ -156,6 +163,9 @@ class SnowGlobeSession:
         "create": {
             "database": _do_ignore,
             None: _do_mutating_noresponse,
+        },
+        "drop": {
+            "table": _do_mutating_noresponse,
         },
         "set": _do_mutating_noresponse,
         "delete": _do_mutating_noresponse,